Add dont-run-too-often logic to clustered tasks
Clustered tasks are now configured with their scheduled period. The cluster lock
service has a new method that doesn't try to get the lock if the mutex was last locked
within the last X milliseconds. Using this feature clustered scheduled tasks are limited
in the rate they execute across the whole cluster